Refactor to pull the ordering of children out into its own module.
The position field lives on in the Mongoid::Tree module mainly so that we can use the :default_order option on the references_many association safely. If I knew a way to make that optional based on whether Mongoid::Tree::Ordering was included, I'd do that. As-is, I don't feel too bad about the unused position field in situations where Mongoid::Tree::Ordering is not included.